pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ag-ui-adk>=0.5.1",
    "cairosvg>=2.7.0",
    "fastapi>=0.135.1",
    "google-adk>=1.26.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.2.2",
    "reportlab>=4.2.5",
    "uvicorn>=0.41.0",
]
